MQQT Producer
To make it look like real-time, this Python script reads temperature data from a CSV file and publishes each entry to an MQTT broker with a 1-second interval. 1 1-second interval is used here for demonstration purposes; in a real-world scenario, this could be adjusted based on actual data frequency. Every time you run this script, it will start publishing from the beginning of the CSV file. It connects to the local MQTT.