IoT warning screen
A small, short video without auditory support. It is designed for an IoT screen to raise an alert to users if the threshold is passed and the contamination of each indicator reaches a dangerous level for the water source. The dashed yellow line is for the acceptable but contaminated range, and the red straight line is for the dangerous range. Crossing the red line causes aquatic organisms to start dying.